chore(*): rename solution
