+++ .travis.yml
@@ -20,7 +20,7 @@ jobs:
         - echo "$DOCKERHUB_PASSWORD" | docker login --username "$DOCKERHUB_LOGIN" --password-stdin
         - docker build . -t fossapps/micro.keystore --build-arg VERSION=$TAG
         - docker-compose -f ./docker-compose.ci.yml up -d
-        - sh ./Micro.Starter.UnitTest/ExternalTests/postman_tests.sh
+        - sh ./Micro.KeyStore.UnitTest/ExternalTests/postman_tests.sh
         - docker tag fossapps/micro.keystore fossapps/micro.keystore:$TAG
         - docker push fossapps/micro.keystore:$TAG
     - stage: deploy
