+++ Micro.KeyStore.Api/Configs/DatabaseConfig.cs
@@ -1,0 +1,11 @@
+namespace Micro.KeyStore.Api.Configs
+{
+    public class DatabaseConfig
+    {
+        public string Host { set; get; }
+        public int Port { set; get; }
+        public string Name { set; get; }
+        public string User { set; get; }
+        public string Password { set; get; }
+    }
+}
