+++ Micro.KeyStore.Api/Configs/SlackLoggingConfig.cs
@@ -1,0 +1,10 @@
+using Microsoft.Extensions.Logging;
+
+namespace Micro.KeyStore.Api.Configs
+{
+    public class SlackLoggingConfig
+    {
+        public string WebhookUrl { set; get; }
+        public LogLevel MinLogLevel { set; get; }
+    }
+}
