+++ Micro.KeyStore.Api/Micro.KeyStore.Api.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="App.Metrics" Version="3.2.0-dev0002" />
+      <PackageReference Include="App.Metrics.AspNetCore.Mvc" Version="3.2.0-dev0002" />
+      <PackageReference Include="App.Metrics.Extensions.Configuration" Version="3.2.0-dev0002" />
+      <PackageReference Include="App.Metrics.Reporting.InfluxDB" Version="3.2.0-dev0002" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.0.0-preview8.19405.11" />
+      <PackageReference Include="Microsoft.Extensions.Logging.Slack" Version="1.1.0" />
+      <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.0.0-preview8" />
+      <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL.Design" Version="1.1.1" />
+      <None Include="./appsettings.ci.json" CopyToPublishDirectory="Always" />
+      <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc2" />
+    </ItemGroup>
+
+    <Target Name="Install githooks" BeforeTargets="Build">
+        <Exec Command="cp ../hooks/* ../.git/hooks/" />
+    </Target>
+</Project>
