+++ Micro.KeyStore.Api/Models/Weather.cs
@@ -1,0 +1,8 @@
+namespace Micro.KeyStore.Api.Models
+{
+    public class Weather
+    {
+        public string Id { set; get; }
+        public double Temperature { set; get; }
+    }
+}
