+++ Micro.KeyStore.Api/Uuid/IUuidService.cs
@@ -1,0 +1,7 @@
+namespace Micro.KeyStore.Api.Uuid
+{
+    public interface IUuidService
+    {
+        string GenerateUuId();
+    }
+}
