+++ Micro.KeyStore.Api/Uuid/UuidService.cs
@@ -1,0 +1,10 @@
+namespace Micro.KeyStore.Api.Uuid
+{
+    public class UuidService : IUuidService
+    {
+        public string GenerateUuId()
+        {
+            return System.Guid.NewGuid().ToString();
+        }
+    }
+}
