+++ Micro.KeyStore.UnitTest/Micro.KeyStore.UnitTest.csproj
