--- Micro.Starter.Api/Configs/DatabaseConfig.cs
@@ -1,11 +1,0 @@
-namespace Micro.Starter.Api.Configs
-{
-    public class DatabaseConfig
-    {
-        public string Host { set; get; }
-        public int Port { set; get; }
-        public string Name { set; get; }
-        public string User { set; get; }
-        public string Password { set; get; }
-    }
-}
