--- Micro.Starter.Api/Configs/SlackLoggingConfig.cs
@@ -1,10 +1,0 @@
-using Microsoft.Extensions.Logging;
-
-namespace Micro.Starter.Api.Configs
-{
-    public class SlackLoggingConfig
-    {
-        public string WebhookUrl { set; get; }
-        public LogLevel MinLogLevel { set; get; }
-    }
-}
