--- Micro.Starter.Api/HealthCheck/HealthData.cs
@@ -1,10 +1,0 @@
-namespace Micro.Starter.Api.HealthCheck
-{
-    public class HealthData
-    {
-        public bool Healthy => FakeDbHealth && FakeCacheHealth;
-
-        public bool FakeDbHealth { set; get; }
-        public bool FakeCacheHealth { set; get; }
-    }
-}
