--- Micro.Starter.Api/Models/Weather.cs
@@ -1,8 +1,0 @@
-namespace Micro.Starter.Api.Models
-{
-    public class Weather
-    {
-        public string Id { set; get; }
-        public double Temperature { set; get; }
-    }
-}
