--- Micro.Starter.Api/Uuid/IUuidService.cs
@@ -1,7 +1,0 @@
-namespace Micro.Starter.Api.Uuid
-{
-    public interface IUuidService
-    {
-        string GenerateUuId();
-    }
-}
