--- Micro.Starter.Api/Uuid/UuidService.cs
@@ -1,10 +1,0 @@
-namespace Micro.Starter.Api.Uuid
-{
-    public class UuidService : IUuidService
-    {
-        public string GenerateUuId()
-        {
-            return System.Guid.NewGuid().ToString();
-        }
-    }
-}
