--- Micro.Starter.UnitTest/UnitTest1.cs
@@ -1,18 +1,0 @@
-using NUnit.Framework;
-
-namespace Tests
-{
-    public class Tests
-    {
-        [SetUp]
-        public void Setup()
-        {
-        }
-
-        [Test]
-        public void Test1()
-        {
-            Assert.Pass();
-        }
-    }
-}
