--- Micro.Starter.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Micro.Starter.Api", "Micro.Starter.Api\Micro.Starter.Api.csproj", "{D782549D-B539-4DA3-B718-D8697EABE4F5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Micro.Starter.UnitTest", "Micro.Starter.UnitTest\Micro.Starter.UnitTest.csproj", "{4384EF12-1CE6-4672-BC61-2F20F83D2744}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D782549D-B539-4DA3-B718-D8697EABE4F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D782549D-B539-4DA3-B718-D8697EABE4F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D782549D-B539-4DA3-B718-D8697EABE4F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D782549D-B539-4DA3-B718-D8697EABE4F5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4384EF12-1CE6-4672-BC61-2F20F83D2744}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4384EF12-1CE6-4672-BC61-2F20F83D2744}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4384EF12-1CE6-4672-BC61-2F20F83D2744}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4384EF12-1CE6-4672-BC61-2F20F83D2744}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
